% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseData.R
\name{browseData}
\alias{browseData}
\title{A browse database function}
\usage{
browseData(dataset = NULL, location = NULL, variables = FALSE,
  collapse = TRUE)
}
\arguments{
\item{dataset}{a character string providing the name of the dataset (optional)}

\item{location}{a character string providing the name of the location (optional)}

\item{variables}{a boolean indicating whether to return the variables of a dataset,
instead of the available locations}

\item{collapse}{boolean indicating whether to return the compact (TRUE) or the extended (FALSE)
data overview, if neither location nor dataset are passed to the function}
}
\value{
\itemize{
   \item if no arguments,  an overview of available data
   \item  if metadata, the requested metadata
   \item  if dataset and variables, available variables for a dataset
   \item  if dataset, avalaible locations for the dataset
   \item  if location, available datasets for a location
   \item  if location and dataset,  retuns a integer. Availability is coded as 0 = no available and 1 = available, This is th quickest option to check availability.
   }
}
\description{
A function to provide information on available data in the
PROFOUND database
}
\details{
Besides providing information on available data, this function allows
to accces the database metadata, policy,  data sources and site descriptions.
}
\note{
To report errors in the package or the data, please use the issue tracker
in the github repository of TG2 https://github.com/COST-FP1304-PROFOUND/TG2/issues
(preferred, but requires that you have access to our GitHub account) or
or use this google form http://goo.gl/forms/e2ZQCiZz4x
}
\examples{
\dontrun{
# See what data is included in the database and for what locations the data is available.
overview <- browseData()
# Hint: If *collapse* FALSE, full version of the overview table
overview <- browseData(collapse = FALSE)

# Available datasets
tables <- browseData(dataset = "DATASETS")

# Available variables for a given dataset
variables <- browseData(dataset = "CLIMATE_LOCAL", variables = TRUE)

# Available locations for a given dataset
available <- browseData(dataset = "CLIMATE_LOCAL")

# Available datasets for a given location
available <- browseData(location ="le_bray")

# Whether a dataset is available for a specific location
available <- browseData(location ="le_bray", dataset = "CLIMATE_LOCAL")

# See version histoy
version <- browseData(dataset = "VERSION")

# Access metadata
metadata <- browseData(dataset = "METADATA_DATASETS")
metadata <- browseData(dataset = "METADATA_CLIMATE_LOCAL")

# Access specific site metadata
metadata <- browseData(dataset = "METADATA_TREE", location = "solling_spruce")

# See data source
source <- browseData(dataset = "SOURCE")

# See specific site data source
source <- browseData(dataset = "SOURCE", location = "solling_spruce")

# See data policy
source <- browseData(dataset = "POLICY")

# See specific site data policy
policy <- browseData(dataset = "POLICY", location = "solling_spruce")
}
}
\keyword{ProfoundData}

