% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{A function to get data}
\usage{
getData(dataset, location = NULL, forcingDataset = NULL,
  forcingCondition = NULL, species = NULL, variables = NULL,
  period = NULL, collapse = TRUE, quality = NULL, decreasing = TRUE)
}
\arguments{
\item{dataset}{a character string providing the name one of the available datasets}

\item{location}{a character string providing the name of the location}

\item{forcingDataset}{a character string providing the name of the a forcingDataset.
Only relevant for ISIMIP datasets.}

\item{forcingCondition}{a character string providing the name of the a forcingCondition.
Only relevant for ISIMIP datasets.}

\item{species}{a character string providing the species name or species id}

\item{variables}{a character array holding the variables to be plotted. Default  is all variables.}

\item{period}{a character array either start of the subset or start and end.
It must have the format YYYY-MM-DD.}

\item{collapse}{a boolean indicating whether the returned data should be a
single dataframe (TRUE) or list of dataframes. Relevant when downloading SOIL and ISIMIP datasets.}

\item{quality}{a number indicating the quality threshold to be used. Default is none.}

\item{decreasing}{a boolean indicating whether the quality threshold should be applied up- or downwards}
}
\value{
a dataframe or list of dataframes, depending on collapse
}
\description{
This function allows to download datasets for a location
from the PROFOUND database.
}
\details{
When using quality, please be aware that the threshold value is included
in the returned data. Threshold works by removing data values taht are greater
 (decreasing = TRUE) or less(decreasing = FALSE) than the given value.
 The quality parameter is only relevant for datasets that have quality flags. These are
ATMOSPHERICHEATCONDUCTION, SOILTS, FLUX, METEOROLOGICAL, and CLIMATE LOCAL for some sites. Please
check the metadata of each dataset before using this parameter.
}
\note{
To report errors in the package or the data, please use the issue tracker
in the github repository of TG2 https://github.com/COST-FP1304-PROFOUND/TG2/issues
(preferred, but requires that you have access to our GitHub account) or
or use this google form http://goo.gl/forms/e2ZQCiZz4x
}
\examples{
\dontrun{
getData("TREE", location = "Soro")
getData("CLIMATE_ISIMIP2B", location = "Soro", collapse=T)
getData("CLIMATE_ISIMIP2B", location = "Soro", collapse=F)
getData("SOIL", location = "Kroof", collapse=F)
getData("SOIL", location = "Kroof", collapse=T)
getData("CLIMATE_LOCAL", location = "Collelongo")
# Period options
getData("CLIMATE_LOCAL", location = "Collelongo", period = "1996-01-01")
getData("CLIMATE_LOCAL", location = "Collelongo", period = c("1996-01-01", "1996-12-31"))
getData("CLIMATE_LOCAL", location = "Collelongo", period = c(NA "1996-12-31"))
getData("CLIMATE_LOCAL", location = "Collelongo", period = c("1996-01-01", NA))
 }
}
\author{
Ramiro Silveyra Gonzalez
}
\keyword{ProfoundData}

