% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDB.R
\name{getDB}
\alias{getDB}
\title{A function to return information of the database connection}
\usage{
getDB()
}
\value{
a character string with the filepath to the database
}
\description{
A function to return the filepath to the database.
}
\examples{
/inst/examples/download-set-get-DBHelp.R
}
\author{
Ramiro Silveyra Gonzalez
}
