% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryDB.R
\name{queryDB}
\alias{queryDB}
\title{A function to create the a database connection object}
\usage{
queryDB(queryItem)
}
\arguments{
\item{queryItem}{a character string providing the query}
}
\description{
A queryDB funtion to  perform own queries to the Profound DB
}
\note{
To report errors in the package or the data, please use the issue tracker
in the github repository of TG2 https://github.com/COST-FP1304-PROFOUND/TG2/issues
(preferred, but requires that you have access to our GitHub account) or
or use this google form http://goo.gl/forms/e2ZQCiZz4x
}
\examples{
\dontrun{
# Basic querz
overview <- queryDB("SELECT * FROM OVERVIEW")
tree <- queryDB("SELECT * FROM TREE")

# Site name or site_id
myQuery <- queryDB("SELECT date, tmax_degC FROM CLIMATE_LOCAL WHERE site == 'hyytiala'")
myQuery <- queryDB("SELECT date, tmax_degC FROM CLIMATE_LOCAL_12")

# Tree species
myQuery <- queryDB("SELECT  * FROM TREE WHERE species == 'Picea abies'")
myQuery <- queryDB("SELECT  * FROM TREE_piab")

# Specify variables
myQuery <- queryDB("SELECT date, tmax_degC FROM CLIMATE_LOCAL WHERE
                   tmax_degC > 20 AND site == 'hyytiala' AND year == 2010")



}
}
\author{
Ramiro Silveyra Gonzalez
}

