% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeData.R
\name{summarizeData}
\alias{summarizeData}
\title{A function to summarize data from the database}
\usage{
summarizeData(dataset, location, forcingDataset = NULL,
  forcingCondition = NULL, by = "year", period = NULL, mode = "data")
}
\arguments{
\item{dataset}{a character string providing the name one of the available
climatic datasets (CLIMATE_LOCAL, CLIMATE_ISIMIP, ...) or the tree dataset}

\item{location}{a character string providing the name of the location}

\item{forcingDataset}{a character string providing the name of the a forcingDataset.
Only relevant for ISIMIP datasets.}

\item{forcingCondition}{a character string providing the name of the a forcingCondition.
Only relevant for ISIMIP datasets.}

\item{by}{a character string indicating how summarize the data. Currently supports
by year and total. The latter refers to the entire available period.}

\item{period}{a character array either start of the subset or start and end.
It must have the format YYYY-MM-DD.}

\item{mode}{a character string whether to display the data summary or an overview.}
}
\value{
a dataframe with the summary values
}
\description{
This function allows to summarize datasets for a location
from the PROFOUND database.
}
\details{
This function is under development and has limited fucntionality. At the
moment, it is possible to summarize daily climate datasets and tree data.

Data are summarized by years. Radiation and precipation are provided as total yearly
values, while the rest of climatic values are year mean values.  For ISIMIP datasets a summary for whole
period will be returned if the dataset comprises more than one forcing datasets and one forcing conditions.
}
\note{
To report errors in the package or the data, please use the issue tracker
in the github repository of TG2 https://github.com/COST-FP1304-PROFOUND/TG2/issues
(preferred, but requires that you have access to our GitHub account) or
or use this google form http://goo.gl/forms/e2ZQCiZz4x
}
\section{Summary values}{
 Summary is calculated by year
\itemize{
  \item Climatic datasets
  \itemize{
   \item  p_mm and rad_Jcm2day are total yearly values
   \item  tmax_degC,  tmean_degC, tmin_degC, relhum_percent, airpress_hPa  and wind_ms are mean yearly values
   }
  \item TREE dataset
   \itemize{
   \item density_treeha is the number of tree per ha
   \item dbhArit_cm is the arithmetic mean diameter
   \item dbhBa_cm is the average diameter weighted by basal area calculated as dbhBA = (ba1*dbh1 + ba2*dbh2 + ... + bak*dbhk) / (ba1 + ba2+ ... + bak), where bai and dbhi are the basal area and dbh, respectively, of the tree i, and  i = 1, 2, . . , k
   \item dbhDQ_cm is the mean squared diameter or quadratic mean diameter calculated as dbhDQ = sqrt( (dbh1^2 + dbh2^2+ ... + dbhk^2) / N),  where dbhi is the diameter at breat height of tree i,   i = 1, 2, . . , k, N is the total number of trees, and sqrt is the square root
   \item heightArith_m is the arithmetic mean height
   \item heightArith_m is the verage height weighted by basal area or Loreys height calculated as heightBA = (ba1*h1 + ba2*h2 + ... + bak*hk) / (ba1 + ba2+ ... + bak), where bai and hi are the basal area and height, respectively, of the tree i, and  i = 1, 2, . . , k
   \item heightBA_m
   \item ba_m2 is the basal area per hectare
   }
}
}
\examples{
\dontrun{
summarizeData("CLIMATE_ISIMIP2B", location = "Soro")
summarizeData("CLIMATE_LOCAL", location = "Soro")
summarizeData("TREE", location = "Soro")
}
}
\author{
Ramiro Silveyra Gonzalez
}
\keyword{ProfoundData}

