% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDB.R
\name{getDB}
\alias{getDB}
\title{A function to retun information the a database connection object}
\usage{
getDB()
}
\value{
a character string with filepath to the database
}
\description{
A getDB function to return the filepath to the database
}
\note{
To report errors in the package or the data, please use the issue tracker
in the github repository of TG2 https://github.com/COST-FP1304-PROFOUND/TG2/issues
(preferred, but requires that you have access to our GitHub account) or
or use this google form http://goo.gl/forms/e2ZQCiZz4x
}
\examples{
\dontrun{
# Retrieve the database connection and version
getDB()
}
}
\author{
Ramiro Silveyra Gonzalez
}

