% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerDatabaseFunctions.R
\name{getDB}
\alias{getDB}
\title{A function to return information of the database connection}
\usage{
getDB()
}
\value{
a character string with the filepath to the database
}
\description{
A function to return the filepath to the database.
}
\examples{
\dontrun{
  
# For downloading the data, use 
dbfile <- downloadDatabase()

# Set the connection using the setDB function
setDB(dbfile)

# check if database is available
getDB()
}
}
\seealso{
\code{\link{setDB}}, \code{\link{downloadDatabase}}
}
\author{
Ramiro Silveyra Gonzalez
}
