% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotData}
\alias{plotData}
\title{A function to plot data}
\usage{
plotData(dataset, location, forcingDataset = NULL, forcingCondition = NULL,
  species = NULL, variables = NULL, period = NULL, aggregated = NULL,
  FUN = mean, automaticPanels = T, quality = NULL, decreasing = TRUE)
}
\arguments{
\item{dataset}{a character string providing the name one of the available datasets}

\item{location}{a character string providing the name of the location}

\item{forcingDataset}{a character string providing the name of the a forcingDataset.
Only relevant for ISIMIP datasets.}

\item{forcingCondition}{a character string providing the name of the a forcingCondition.
Only relevant for ISIMIP datasets.}

\item{species}{a character string providing the species name or species id.}

\item{variables}{a character array holding the variables to be plotted. Default  is all variables.}

\item{period}{a character array either start of the subset or start and end}

\item{aggregated}{a boolean indicating whether data should be aggregated for display.
Possible values are date, day, month and year.}

\item{FUN}{a function to use for aggregating the data}

\item{automaticPanels}{should the function automatically create panels}

\item{quality}{a number indicating the quality threshold to be used. Default is none.}

\item{decreasing}{a boolean indicating whether the quality threshold should be applied up- or downwards}
}
\value{
plots for the specified dataset, location and variables
}
\description{
A function to plot data of dataset for a location
from the ProfoundData database.
}
\details{
Plotting is not supported for datasets: OVERVIEW, SITES, SOIL. The aggregation of data
relies on the function \code{\link{aggregate}} for \code{\link{zoo}} objects. The FUN parameter
is passed to FUN from \code{\link{aggregate}}. Please check the help files of  \code{\link{aggregate}}
for further information. For handling NAs we recommend to pass self-defined fuctions (see examples)
the behavious
}
\note{
To report errors in the package or the data, please use the issue tracker
in the github repository of TG2 https://github.com/COST-FP1304-PROFOUND/TG2/issues
(preferred, but requires that you have access to our GitHub account) or
or use this google form http://goo.gl/forms/e2ZQCiZz4x
}
\examples{
\dontrun{
plotData("TREE", location = "Soro")
plotData("TREE", location = "Soro", automaticPanels=T)
plotData("TREE", location = "Soro", automaticPanels=T, aggregate = T, FUN= mean)
plotData("TREE", location = "Soro", automaticPanels=T, FUN = function(x)mean(x, na.rm=T))
#' # Period options
plotData("CLIMATE_LOCAL", location = "Collelongo", period = "1996-01-01")
plotData("CLIMATE_LOCAL", location = "Collelongo", period = c("1996-01-01", "1996-12-31"))
plotData("CLIMATE_LOCAL", location = "Collelongo", period = c(NA "1996-12-31"))
plotData("CLIMATE_LOCAL", location = "Collelongo", period = c("1996-01-01", NA))
 }
}
\author{
Ramiro Silveyra Gonzalez
}
\keyword{ProfoundData}

