% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportDB.R
\name{reportDB}
\alias{reportDB}
\title{A function report data}
\usage{
reportDB(outDir = NULL)
}
\arguments{
\item{outDir}{a character string indicating the output directory in which the
html file will be saved. If no value is provided, the working directory will be
used as output directory.}
}
\value{
a html file with the database report
}
\description{
This functions creates a site-by-site report of all avalaible data
in the PROFOUND database. The summary is created with a rmarkdown document, which
is rendered and saved as a html document.
}
\details{
Please note that creating the report it might take several minutes.
}
\note{
To report errors in the package or the data, please use the issue tracker
in the GitHub repository of ProfoundData \url{https://github.com/COST-FP1304-PROFOUND/ProfoundData}
}
\examples{
# example requires that a sql DB is registered via setDB(dbfile)
# when run without a registered DB, you will get a file query (depending on OS)

\dontrun{
# Create the database report and specify the output directory
reportDB(outDir = "/home/mine/database/")
  
# Create the database report without specify the output directory
reportDB()
}
}
