% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDB.R
\name{setDB}
\alias{setDB}
\title{A function to set the connection to the database}
\usage{
setDB(db_name)
}
\arguments{
\item{db_name}{a character string providing the absolute path to the PROFOUND database.}
}
\description{
A setDB funtion to  create a database connection object
}
\note{
To report errors in the package or the data, please use the issue tracker
in the github repository of TG2 https://github.com/COST-FP1304-PROFOUND/TG2/issues
(preferred, but requires that you have access to our GitHub account) or
or use this google form http://goo.gl/forms/e2ZQCiZz4x
}
\examples{
\dontrun{
# Define the path to the database. It must be an absolute path.
myDB <- "/home/ownCloud/PROFOUND_Data/v0.1.13/ProfoundData.sqlite"
myDB <- path.expand("~/ownCloud/PROFOUND_Data/v0.1.13/ProfoundData.sqlite")

# Set the connection using the setDB function
setDB(myDB)
}
}
\author{
Ramiro Silveyra Gonzalez
}

