% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDB.R
\name{setDB}
\alias{setDB}
\title{A function to create the a database connection object}
\usage{
setDB(db_name)
}
\arguments{
\item{db_name}{a character string providing the absolute path to the Profound DB.}
}
\description{
A setDB funtion to  create a database connection object
}
\note{
To report errors in the package or the data, please use the issue tracker
in the github repository of TG2 https://github.com/COST-FP1304-PROFOUND/TG2/issues
(preferred, but requires that you have access to our GitHub account) or
or use this google form http://goo.gl/forms/e2ZQCiZz4x
}
\examples{
 \dontrun{
  setDB(db_name = "~/PROFOUND_Data/ProfoundData.sqlite")
 }
}
\author{
Ramiro Silveyra Gonzalez
}

