% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerDatabaseFunctions.R
\name{setDB}
\alias{setDB}
\title{A function to set the connection to the database}
\usage{
setDB(db_name = NULL)
}
\arguments{
\item{db_name}{a character string providing the absolute path to the PROFOUND database.}
}
\description{
A setDB funtion to  create a database connection object
}
\note{
To report errors in the package or the data, please use the issue tracker
in the GitHub repository of ProfoundData \url{https://github.com/COST-FP1304-PROFOUND/ProfoundData}
}
\examples{
\dontrun{
  
# For downloading the data, use 
downloadDatabase()
# The function will query you to ask about the path to store the databse
  
# Define the path to the database. 
myDB <- "/home/ownCloud/PROFOUND_Data/v0.1.13/ProfoundData.sqlite"
# myDB must be an absolute paths, so you should expand relative paths
myDB <- path.expand("~/ownCloud/PROFOUND_Data/v0.1.13/ProfoundData.sqlite")

# Set the connection using the setDB function
setDB(myDB)

# check if database is available
getDB()
}
}
